Expand all transitions from the .sgf file
In the .sgf grammar it is possible to leave out source and destination,
if they repeat. The definition then applies to the previous source
and/or destination. In the final state machine definition every
transition should exist stand alone. Hence all transitions are expanded
with the information of the previous transitions via the function
`expandTransitions`.