Simplify the parser
The parser state now only stores the values required for the actual
parsing: startcode, current input and previous token.

This simplifies the parser code, but the result of the generated parser
is only partially made up of the desired data types. Subsequent compiler
runs will have to achieve the final result.